Style of the pages and Navbar
Two new files with names thanks and contact created.

The navbar created to all pages, but there is a little bit problem with connecting them, do not know why we can't follow links in website. It needs review.

By using a form and a map, created the contact page.

In the about page I have just used a picture with a sample text. So if we think this is not enough we can change it.